add route and GetX
