+++ .dart_tool/flutter_build/d7c8b0d8d7d9c86f968efb8a90aefabf/outputs.json
@@ -1,1 +1,1 @@
-["/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/vm_snapshot_data","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/isolate_snapshot_data","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/kernel_blob.bin","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/assets/images/logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/assets/images/asp.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/cupertino_icons/assets/CupertinoIcons.ttf","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/fonts/MaterialIcons-Regular.otf","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/shaders/ink_sparkle.frag","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/AssetManifest.json","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/AssetManifest.bin","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/FontManifest.json","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/NOTICES.Z"]
+["/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/vm_snapshot_data","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/isolate_snapshot_data","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/kernel_blob.bin","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/assets/images/logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/assets/images/asp.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/cupertino_icons/assets/CupertinoIcons.ttf","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/google-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/2.0x/google-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/3.0x/google-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/facebook-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/2.0x/facebook-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/3.0x/facebook-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/apple-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/2.0x/apple-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/3.0x/apple-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/apple-black-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/2.0x/apple-black-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/3.0x/apple-black-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/github-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/2.0x/github-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/3.0x/github-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/microsoft-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/2.0x/microsoft-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/3.0x/microsoft-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/twitter-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/2.0x/twitter-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/packages/social_login_buttons/lib/assets/3.0x/twitter-logo.png","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/fonts/MaterialIcons-Regular.otf","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/shaders/ink_sparkle.frag","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/AssetManifest.json","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/AssetManifest.bin","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/FontManifest.json","/Users/asepherbasuki/Documents/Altea/project/flutter/flutter_bootcamp_tugas1/build/app/intermediates/flutter/debug/flutter_assets/NOTICES.Z"]
