+++ .dart_tool/package_config.json
@@ -62,6 +62,12 @@
       "languageVersion": "3.0"
     },
     {
+      "name": "get",
+      "rootUri": "file:///Users/asepherbasuki/.pub-cache/hosted/pub.dev/get-4.6.5",
+      "packageUri": "lib/",
+      "languageVersion": "2.15"
+    },
+    {
       "name": "js",
       "rootUri": "file:///Users/asepherbasuki/.pub-cache/hosted/pub.dev/js-0.6.7",
       "packageUri": "lib/",
@@ -102,6 +108,12 @@
       "rootUri": "file:///usr/local/Caskroom/flutter/2.10.5/flutter/bin/cache/pkg/sky_engine",
       "packageUri": "lib/",
       "languageVersion": "3.0"
+    },
+    {
+      "name": "social_login_buttons",
+      "rootUri": "file:///Users/asepherbasuki/.pub-cache/hosted/pub.dev/social_login_buttons-1.0.7",
+      "packageUri": "lib/",
+      "languageVersion": "2.12"
     },
     {
       "name": "source_span",
@@ -152,7 +164,7 @@
       "languageVersion": "3.0"
     }
   ],
-  "generated": "2023-05-22T10:47:59.095366Z",
+  "generated": "2023-05-23T07:33:02.688110Z",
   "generator": "pub",
   "generatorVersion": "3.0.0"
 }
