+++ .dart_tool/package_config_subset
@@ -30,6 +30,10 @@ flutter_lints
 2.17
 file:///Users/asepherbasuki/.pub-cache/hosted/pub.dev/flutter_lints-2.0.1/
 file:///Users/asepherbasuki/.pub-cache/hosted/pub.dev/flutter_lints-2.0.1/lib/
+get
+2.15
+file:///Users/asepherbasuki/.pub-cache/hosted/pub.dev/get-4.6.5/
+file:///Users/asepherbasuki/.pub-cache/hosted/pub.dev/get-4.6.5/lib/
 js
 2.19
 file:///Users/asepherbasuki/.pub-cache/hosted/pub.dev/js-0.6.7/
@@ -54,6 +58,10 @@ path
 2.12
 file:///Users/asepherbasuki/.pub-cache/hosted/pub.dev/path-1.8.3/
 file:///Users/asepherbasuki/.pub-cache/hosted/pub.dev/path-1.8.3/lib/
+social_login_buttons
+2.12
+file:///Users/asepherbasuki/.pub-cache/hosted/pub.dev/social_login_buttons-1.0.7/
+file:///Users/asepherbasuki/.pub-cache/hosted/pub.dev/social_login_buttons-1.0.7/lib/
 source_span
 2.14
 file:///Users/asepherbasuki/.pub-cache/hosted/pub.dev/source_span-1.9.1/
