+++ .gitignore
@@ -255,3 +255,8 @@ GoogleService-Info.plist
 google-services.json
 # Plugins (already resolved through pubspec.yaml)
 .flutter-plugins
+
+# Android Studio will place build artifacts here
+/android/app/debug
+/android/app/profile
+/android/app/releases
