+++ lib/constants/default_constants.dart
@@ -1,0 +1,9 @@
+// ignore_for_file: non_constant_identifier_names, constant_identifier_names
+
+abstract class NamesConstans {
+  NamesConstans._();
+
+  static const NEW_USER = "New User? Create Account";
+  static const APPS_TITLE = "Bootcamp 14 : Flutter Fundamental";
+
+}
