+++ lib/main.dart
@@ -1,94 +1,24 @@
 import 'package:flutter/material.dart';
-import 'home_page.dart';
+import 'package:flutter_bootcamp_tugas1_widget/constants/default_constants.dart';
+import 'package:get/get.dart';
+import 'package:get/route_manager.dart';
+import 'routes/routes_list.dart';
 
-void main() {
-  runApp(LoginScreenPage());
+
+void main() async {
+  runApp(const MyApp());
 }
 
-class LoginScreenPage extends StatelessWidget {
-  const LoginScreenPage({super.key});
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return GetMaterialApp(
+      title: NamesConstans.APPS_TITLE,
       debugShowCheckedModeBanner: false,
-      home: LoginDemo(),
-    );
-  }
-}
-
-class LoginDemo extends StatefulWidget {
-  @override
-  _LoginDemoState createState() => _LoginDemoState();
-}
-
-class _LoginDemoState extends State<LoginDemo> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,  
-      appBar: AppBar(
-        title: const Text("Login Page"),
-      ),
-      body: SingleChildScrollView(
-        child: Column(
-          children: <Widget>[
-            Padding(
-              padding: const EdgeInsets.only(top: 60.0),
-              child: Center(
-                child: SizedBox(
-                    width: 200,
-                    height: 150,
-                    child: Image.asset('assets/images/logo.png')),
-              ),
-            ),
-            const Padding(
-              padding: EdgeInsets.symmetric(horizontal: 15),
-              child: TextField(
-                decoration: InputDecoration(
-                    border: OutlineInputBorder(),
-                    labelText: 'Email',
-                    hintText: 'Enter valid email'),
-              ),
-            ),
-            const Padding(
-              padding: EdgeInsets.only(
-                  left: 15.0, right: 15.0, top: 15, bottom: 0),
-              child: TextField(
-
-                obscureText: true,
-                decoration: InputDecoration(
-                    border: OutlineInputBorder(),
-                    labelText: 'Password',
-                    hintText: 'Enter password'),
-              ),
-            ),
-            TextButton(onPressed: (){}, child: const Text(
-                'Forgot Password',
-                style: TextStyle(color: Colors.blue, fontSize: 15),
-              ),
-            ),
-            Container(
-              height: 50,
-              width: 250,
-              decoration: BoxDecoration(
-                  color: Colors.blue, borderRadius: BorderRadius.circular(20)),
-              child: TextButton(
-                onPressed: () {
-                  Navigator.push(
-                      context, MaterialPageRoute(builder: (_) => const HomePage()));
-                },
-                child: const Text(
-                  'Login',
-                  style: TextStyle(color: Colors.white, fontSize: 25),
-                ),
-              ),
-            ),
-            const SizedBox(height: 130),
-            const Text('New User? Create Account')
-          ],
-        ),
-      ),
+      getPages: MyRoutes.routes,
+      initialRoute: MyRoutes.getLoginRoute(),
     );
   }
 }
