+++ lib/pages/login/login_page.dart
@@ -1,0 +1,99 @@
+import 'package:flutter/material.dart';
+import 'package:social_login_buttons/social_login_buttons.dart';
+
+import '../home/home_page.dart';
+
+class LoginPage extends StatefulWidget {
+  const LoginPage({Key? key}) : super(key: key);
+
+  @override
+  State<LoginPage> createState() => _LoginPageState();
+}
+
+class _LoginPageState extends State<LoginPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      // backgroundColor: Colors.white,  
+      // appBar: AppBar(
+      //   title: const Text("Login Page"),
+      // ),
+      body: SingleChildScrollView(
+        child: Column(
+          children: <Widget>[
+            Padding(
+              padding: const EdgeInsets.only(top: 60.0),
+              child: Center(
+                child: SizedBox(
+                    width: 200,
+                    height: 150,
+                    child: Image.asset('assets/images/logo.png')),
+              ),
+            ),
+            const Padding(
+              padding: EdgeInsets.symmetric(horizontal: 15),
+              child: TextField(
+                decoration: InputDecoration(
+                    border: OutlineInputBorder(),
+                    labelText: 'Email',
+                    hintText: 'Enter valid email'),
+              ),
+            ),
+            const Padding(
+              padding: EdgeInsets.only(
+                  left: 15.0, right: 15.0, top: 15, bottom: 0),
+              child: TextField(
+
+                obscureText: true,
+                decoration: InputDecoration(
+                    border: OutlineInputBorder(),
+                    labelText: 'Password',
+                    hintText: 'Enter password'),
+              ),
+            ),
+            TextButton(onPressed: (){}, child: const Text(
+                'Forgot Password',
+                style: TextStyle(color: Colors.blue, fontSize: 15),
+              ),
+            ),
+            Container(
+              height: 50,
+              width: 250,
+              decoration: BoxDecoration(
+                  color: Colors.blue, borderRadius: BorderRadius.circular(20)),
+              child: TextButton(
+                onPressed: () {
+                  Navigator.push(
+                      context, MaterialPageRoute(builder: (_) => const HomePage()));
+                },
+                child: const Text(
+                  'Login',
+                  style: TextStyle(color: Colors.white, fontSize: 25),
+                ),
+              ),
+            ),
+            const SizedBox(height: 25),
+            // const Text(NamesConstans.NEW_USER),
+            // const SizedBox(
+            //   height: 28.0,
+            // ),
+            const Text(
+              '-OR-',
+               style: TextStyle(
+                fontSize: 18.0,
+                color: Color.fromARGB(255, 9, 9, 9),
+                fontWeight: FontWeight.bold),
+            ),
+            const SizedBox(
+              height: 23.0,
+            ),
+            SocialLoginButton(
+              buttonType: SocialLoginButtonType.google,
+              onPressed: () {},
+            ),    
+          ],
+        ),
+      ),
+    );
+  }
+}
